Add files via upload
In my first project, I created a simple calculator using the C programming language. The calculator allows the user to perform basic arithmetic operations such as addition, subtraction, multiplication, and division. The program prompts the user to enter an operator and two numbers, and then performs the corresponding operation based on the operator chosen. The result is displayed on the screen